+++ backend/src/main.go
@@ -7,12 +7,10 @@ import (
 	"log"
 	"net/http"
 	"os"
-	"regexp"
 	"strconv"
 	"time"
 
 	"github.com/waterproofpatch/go_authentication/authentication"
-	"gorm.io/gorm"
 
 	"github.com/gorilla/handlers"
 	"github.com/gorilla/mux"
@@ -43,68 +41,6 @@ func startServing(port int, router *mux.Router) {
 	}
 
 	log.Fatal(srv.ListenAndServe())
-}
-
-func startTimer(stopCh chan bool, db *gorm.DB) {
-	// Create a ticker that ticks every 5 seconds
-	ticker := time.NewTicker(5 * time.Second)
-
-	for {
-		select {
-		case <-ticker.C:
-			fmt.Println("Function executed at", time.Now())
-			var plants []app.PlantModel
-			db.Find(&plants)
-			for _, plant := range plants {
-				fmt.Printf("checking watering notifications for %v", plant)
-				regex := regexp.MustCompile(`\s*\([^)]*\)`)
-				lastWaterDateStr := regex.ReplaceAllString(plant.LastWaterDate, "")
-				notifyDateStr := regex.ReplaceAllString(plant.LastNotifyDate, "")
-				lastWaterDateLayout := "Mon Jan 02 2006 15:04:05 MST-0700"
-				lastNotifyDateLayout := "2006-01-02 15:04:05.999999999 -0700 MST"
-				lastWaterDate, err := time.Parse(lastWaterDateLayout, lastWaterDateStr)
-				if err != nil {
-					fmt.Println("Error parsing lastWaterDate string:", err)
-					break
-				}
-				notifyDate, err := time.Parse(lastNotifyDateLayout, notifyDateStr)
-				if err != nil {
-					fmt.Println("Error parsing lastNotifyDate string:", err)
-					break
-				}
-
-				wateringFrequency, err := strconv.Atoi(plant.WateringFrequency)
-				if err != nil {
-					fmt.Println("Error converting string to integer:", err)
-					return
-				}
-
-				nextWaterDate := lastWaterDate.AddDate(0, 0, wateringFrequency)
-				fmt.Printf("Next watering date: %v\n", nextWaterDate)
-				today := time.Now().UTC()
-				if nextWaterDate.Before(today) {
-					fmt.Printf("next water date is before today=%v\n", nextWaterDate)
-					currentDate := time.Now().UTC()
-					diff := currentDate.Sub(notifyDate).Hours()
-					fmt.Printf("lastNotifyDate=%v", notifyDate)
-					fmt.Printf("It's been %v hours since last notification\n", diff)
-					if diff > 24 {
-						fmt.Printf("notify!\n")
-						plant.LastNotifyDate = currentDate.String()
-						db.Save(&plant)
-					}
-
-				}
-
-			}
-			// for each plant, check if it's time to water it
-		case <-stopCh:
-			// Stop the ticker and exit the goroutine
-			fmt.Println("Stopping timer...")
-			ticker.Stop()
-			return
-		}
-	}
 }
 
 // main is the entrypoint to the program.
